Reorganize prompts into templates/<feature> submodules
Split app/prompts/{ats,enrichment,refinement,templates}.py into
app/prompts/templates/<feature>/<prompt>.py modules and rename every
exported identifier with a feature prefix for clearer call-site reads.
Remove the legacy markdown-fallback improve_resume(); the improve and
preview endpoints now return 422 when a resume lacks structured data.
