+++ apps/backend/app/prompts/templates/__init__.py
@@ -1,0 +1,5 @@
+"""LLM prompt templates, grouped by feature module."""
+
+from app.prompts.templates.util import LANGUAGE_NAMES, get_language_name
+
+__all__ = ["LANGUAGE_NAMES", "get_language_name"]
