+++ apps/backend/app/prompts/templates/ats/__init__.py
@@ -1,0 +1,6 @@
+"""ATS scoring and optimization prompts."""
+
+from app.prompts.templates.ats.optimize import ATS_OPTIMIZE_PROMPT
+from app.prompts.templates.ats.score import ATS_SCORE_PROMPT
+
+__all__ = ["ATS_OPTIMIZE_PROMPT", "ATS_SCORE_PROMPT"]
