+++ apps/backend/app/prompts/templates/cover_letter/__init__.py
@@ -1,0 +1,9 @@
+"""Cover letter and outreach message prompts."""
+
+from app.prompts.templates.cover_letter.cover_letter import COVER_LETTER_PROMPT
+from app.prompts.templates.cover_letter.outreach import COVER_LETTER_OUTREACH_PROMPT
+
+__all__ = [
+    "COVER_LETTER_OUTREACH_PROMPT",
+    "COVER_LETTER_PROMPT",
+]
