+++ apps/backend/app/prompts/templates/cover_letter/outreach.py
@@ -1,0 +1,23 @@
+"""Cold outreach message prompt."""
+
+COVER_LETTER_OUTREACH_PROMPT = """Generate a cold outreach message for LinkedIn or email about this job opportunity.
+
+IMPORTANT: Write in {output_language}.
+
+Job Description:
+{job_description}
+
+Candidate Resume (JSON):
+{resume_data}
+
+Guidelines:
+- 70-100 words maximum (shorter than a cover letter)
+- First sentence: Reference specific detail from job description (team, product, technical challenge) - never open with "I'm reaching out" or "I saw your posting"
+- One sentence on strongest matching qualification with a concrete metric if available
+- End with low-friction ask: "Worth a quick chat?" not "I'd love the opportunity to discuss"
+- Tone: How you'd message a former colleague, not a stranger
+- Do NOT include placeholder brackets
+- Do NOT use phrases like "excited about" or "passionate about"
+- Do NOT use em dash ("—") anywhere in the writing/output, even if it exists, remove it
+
+Output plain text only. No JSON, no markdown formatting."""
