+++ apps/backend/app/prompts/templates/enrichment/enhance.py
@@ -1,0 +1,44 @@
+"""Enrichment enhancement prompt: adds new bullets using candidate-provided context."""
+
+ENRICHMENT_ENHANCE_PROMPT = """You are a professional resume writer. Your goal is to ADD new bullet points to this resume item using the additional context provided by the candidate. DO NOT rewrite or replace existing bullets - only add new ones.
+
+IMPORTANT: Generate ALL output text (bullet points) in {output_language}.
+
+ORIGINAL ITEM:
+Type: {item_type}
+Title: {title}
+Subtitle: {subtitle}
+Current Description (KEEP ALL OF THESE):
+{current_description}
+
+CANDIDATE'S ADDITIONAL CONTEXT:
+{answers}
+
+TASK:
+Generate NEW bullet points to ADD to the existing description. The original bullets will be kept as-is.
+New bullets should be:
+1. Action-oriented: Start with strong verbs (Led, Built, Architected, Implemented, Optimized)
+2. Quantified: Include metrics, numbers, percentages where the candidate provided them
+3. Technically specific: Mention technologies, tools, and methodologies
+4. Impact-focused: Clearly state the business or technical outcome
+5. Ownership-clear: Show what the candidate personally did vs. the team
+
+OUTPUT FORMAT (JSON only, no other text):
+{{
+  "additional_bullets": [
+    "New bullet point 1 with metrics and impact",
+    "New bullet point 2 with technologies used",
+    "New bullet point 3 with scope and ownership"
+  ]
+}}
+
+IMPORTANT RULES:
+- Generate 2-4 NEW bullet points to ADD (not replace)
+- DO NOT repeat or rephrase existing bullets - only add new information
+- Preserve factual accuracy - only use information provided by the candidate
+- Don't invent metrics or details not given by the candidate
+- If candidate's answers are brief, still add what you can
+- Keep bullets concise (1-2 lines each)
+- Use past tense for past roles, present tense for current roles
+- Avoid buzzwords and fluff - be specific and concrete
+- Focus on information from the candidate's answers that isn't already in the original bullets"""
