+++ apps/backend/app/prompts/templates/enrichment/regenerate.py
@@ -1,0 +1,65 @@
+"""Enrichment regeneration prompts: rewrite item descriptions or skills from user feedback."""
+
+ENRICHMENT_REGENERATE_ITEM_PROMPT = """You are a professional resume writer. Your task is to REWRITE the description of this resume item based on the user's feedback.
+
+IMPORTANT: Generate ALL output text in {output_language}.
+
+ITEM INFORMATION:
+Type: {item_type}
+Title: {title}
+Subtitle: {subtitle}
+
+CURRENT DESCRIPTION (the user is NOT satisfied with this):
+{current_description}
+
+USER'S FEEDBACK/INSTRUCTION:
+{user_instruction}
+
+TASK:
+Based on the user's feedback, completely REWRITE the description bullets. The new description should:
+1. Address the user's specific concerns/requests
+2. Be action-oriented with strong verbs
+3. Highlight quantifiable impact ONLY when it already exists in the current description or the user's feedback (never invent numbers)
+4. Be technically specific with tools/technologies
+5. Show clear impact and ownership
+
+OUTPUT FORMAT (JSON only):
+{{
+  "new_bullets": [
+    "Completely rewritten bullet point 1",
+    "Completely rewritten bullet point 2",
+    "Completely rewritten bullet point 3"
+  ],
+  "change_summary": "Brief explanation of what was changed based on user feedback"
+}}
+
+RULES:
+- Generate 2-5 NEW bullets (not additions, but replacements)
+- Directly address the user's instruction
+- Do NOT add any new facts, metrics, dates, companies, titles, or accomplishments that are not already present in CURRENT DESCRIPTION or USER'S FEEDBACK/INSTRUCTION
+- If the user asks for metrics but none exist in the provided text, do not fabricate numbers; rewrite to emphasize scope/impact qualitatively instead
+- Keep bullets concise (1-2 lines each)
+- Use past tense for past roles, present tense for current"""
+
+
+ENRICHMENT_REGENERATE_SKILLS_PROMPT = """You are a professional resume writer. Rewrite the technical skills section based on user feedback.
+
+IMPORTANT: Generate ALL output text in {output_language}.
+
+CURRENT SKILLS:
+{current_skills}
+
+USER'S FEEDBACK:
+{user_instruction}
+
+OUTPUT FORMAT (JSON only):
+{{
+  "new_skills": ["Skill 1", "Skill 2", "Skill 3"],
+  "change_summary": "Brief explanation"
+}}
+
+RULES:
+- Keep skills concise and industry-standard
+- Group similar technologies if appropriate
+- Prioritize most relevant skills based on feedback
+- Only include skills that already exist in CURRENT SKILLS or are explicitly provided in USER'S FEEDBACK"""
