+++ apps/backend/app/prompts/templates/job_description/keywords.py
@@ -1,0 +1,20 @@
+"""Keyword extraction prompt — pulls required skills, responsibilities, etc. from a JD."""
+
+JOB_DESCRIPTION_EXTRACT_KEYWORDS_PROMPT = """Extract job requirements as JSON. Output ONLY the JSON object, no other text.
+
+Example format:
+{{
+  "required_skills": ["Python", "AWS"],
+  "preferred_skills": ["Kubernetes"],
+  "experience_requirements": ["5+ years"],
+  "education_requirements": ["Bachelor's in CS"],
+  "key_responsibilities": ["Lead team"],
+  "keywords": ["microservices", "agile"],
+  "experience_years": 5,
+  "seniority_level": "senior"
+}}
+
+Extract numeric years (e.g., "5+ years" → 5) and infer seniority level.
+
+Job description:
+{job_description}"""
