+++ apps/backend/app/prompts/templates/job_description/title.py
@@ -1,0 +1,17 @@
+"""Job title extraction prompt (parses role + company from a JD)."""
+
+JOB_DESCRIPTION_TITLE_PROMPT = """Extract the job title and company name from this job description.
+
+IMPORTANT: Write in {output_language}.
+
+Job Description:
+{job_description}
+
+Rules:
+- Format: "Role @ Company" (e.g., "Senior Frontend Engineer @ Stripe")
+- If the company name is not found, return just the role (e.g., "Senior Frontend Engineer")
+- Maximum 60 characters
+- Use the most specific role title mentioned
+- Do not add any other text, quotes, or formatting
+
+Output the title only, nothing else."""
