+++ apps/backend/app/prompts/templates/resume/improve.py
@@ -1,0 +1,80 @@
+"""Keyword-enhancement diff prompt.
+
+Used when the user picks "Keyword enhance" mode in the improve flow. The prompt
+weaves JD keywords into existing bullets where the resume already supports them
+and may rephrase bullets, but does not invent new ones. The other two strategies
+have their own dedicated prompts: ``nudge.py`` (most conservative) and
+``full_tailor.py`` (most aggressive).
+"""
+
+RESUME_IMPROVE_PROMPT ="""Given this resume and job description, output a JSON object with targeted changes to better align the resume with the job.
+
+RULES:
+1. Only modify content — never change names, companies, dates, institutions, or degrees
+2. Do not invent metrics or achievements not supported by the original resume text
+3. Do not add new work entries, education entries, or project entries
+4. Weave in relevant keywords where evidence already exists. You may rephrase bullets but do not add new ones.
+5. Each change MUST include the original text (copied exactly) so it can be verified
+6. For each change, explain WHY it helps match the job description
+7. Generate all new text in {output_language}
+8. Do not use em dash characters
+9. Keep changes minimal and targeted — do not rewrite content that already aligns well
+10. Exception to rule 2: you may add a skill only if it appears in the verified skill targets below
+11. Improve work and project bullets around the verified skill targets when the original text supports that alignment
+
+PATHS you can target:
+- "summary" — the resume summary text
+- "workExperience[i].description[j]" — a specific bullet (i = entry index, j = bullet index)
+- "workExperience[i].description" — append a new bullet (action: "append")
+- "personalProjects[i].description[j]" — a specific project bullet
+- "personalProjects[i].description" — append a new project bullet
+- "additional.technicalSkills" — reorder the skills list (action: "reorder") or add one verified skill (action: "add_skill")
+
+Do NOT target: personalInfo, dates/years, company names, education, customSections.
+
+Keywords to emphasize (only if already supported by resume content):
+{job_keywords}
+
+Verified skill targets:
+{skill_targets}
+
+Job Description:
+{job_description}
+
+Original Resume:
+{original_resume}
+
+Output this exact JSON format, nothing else:
+{{
+  "changes": [
+    {{
+      "path": "workExperience[0].description[1]",
+      "action": "replace",
+      "original": "the exact original text at this path",
+      "value": "the improved text",
+      "reason": "why this change helps"
+    }},
+    {{
+      "path": "summary",
+      "action": "replace",
+      "original": "the current summary text",
+      "value": "the improved summary",
+      "reason": "why this change helps"
+    }},
+    {{
+      "path": "additional.technicalSkills",
+      "action": "reorder",
+      "original": null,
+      "value": ["most relevant skill first", "then next", "..."],
+      "reason": "reordered to prioritize JD-relevant skills"
+    }},
+    {{
+      "path": "additional.technicalSkills",
+      "action": "add_skill",
+      "original": null,
+      "value": "verified skill target missing from the skills list",
+      "reason": "added verified JD skill for review"
+    }}
+  ],
+  "strategy_notes": "brief summary of the tailoring approach"
+}}"""
