+++ apps/backend/app/prompts/templates/resume/nudge.py
@@ -1,0 +1,61 @@
+"""Light-nudge diff prompt.
+
+Used when the user picks "Light nudge" mode in the improve flow — the most
+conservative strategy. The prompt only allows rephrasing existing bullets where
+the resume already has a clear match for the JD; it never adds bullets.
+"""
+
+RESUME_NUDGE_PROMPT = """Given this resume and job description, output a JSON object with targeted changes to better align the resume with the job.
+
+RULES:
+1. Only modify content — never change names, companies, dates, institutions, or degrees
+2. Do not invent metrics or achievements not supported by the original resume text
+3. Do not add new work entries, education entries, or project entries
+4. Make minimal edits. Only rephrase where there is a clear match. Do not add new bullet points.
+5. Each change MUST include the original text (copied exactly) so it can be verified
+6. For each change, explain WHY it helps match the job description
+7. Generate all new text in {output_language}
+8. Do not use em dash characters
+9. Keep changes minimal and targeted — do not rewrite content that already aligns well
+10. Exception to rule 2: you may add a skill only if it appears in the verified skill targets below
+
+PATHS you can target:
+- "summary" — the resume summary text
+- "workExperience[i].description[j]" — a specific bullet (i = entry index, j = bullet index)
+- "personalProjects[i].description[j]" — a specific project bullet
+- "additional.technicalSkills" — reorder the skills list (action: "reorder") or add one verified skill (action: "add_skill")
+
+Do NOT target: personalInfo, dates/years, company names, education, customSections.
+
+Keywords to emphasize (only if already supported by resume content):
+{job_keywords}
+
+Verified skill targets:
+{skill_targets}
+
+Job Description:
+{job_description}
+
+Original Resume:
+{original_resume}
+
+Output this exact JSON format, nothing else:
+{{
+  "changes": [
+    {{
+      "path": "workExperience[0].description[1]",
+      "action": "replace",
+      "original": "the exact original text at this path",
+      "value": "the rephrased text",
+      "reason": "why this change helps"
+    }},
+    {{
+      "path": "additional.technicalSkills",
+      "action": "reorder",
+      "original": null,
+      "value": ["most relevant skill first", "then next", "..."],
+      "reason": "reordered to prioritize JD-relevant skills"
+    }}
+  ],
+  "strategy_notes": "brief summary of the tailoring approach"
+}}"""
