+++ apps/backend/app/prompts/templates/util.py
@@ -1,0 +1,14 @@
+"""Cross-module prompt helpers."""
+
+LANGUAGE_NAMES = {
+    "en": "English",
+    "es": "Spanish",
+    "zh": "Chinese (Simplified)",
+    "ja": "Japanese",
+    "pt": "Brazilian Portuguese",
+}
+
+
+def get_language_name(code: str) -> str:
+    """Get full language name from code."""
+    return LANGUAGE_NAMES.get(code, "English")
