+++ apps/backend/app/services/__init__.py
@@ -1,13 +1,12 @@
 """Business logic services."""
 
 from app.services.parser import parse_document, parse_resume_to_json
-from app.services.improver import improve_resume, generate_improvements
+from app.services.improver import generate_improvements
 from app.services.refiner import refine_resume
 
 __all__ = [
     "parse_document",
     "parse_resume_to_json",
-    "improve_resume",
     "generate_improvements",
     "refine_resume",
 ]
